+++ dsmr_frontend/migrations/0006_notifications_initial.py
@@ -23,7 +23,7 @@ def insert_notifications(apps, schema_editor):
         message=dsmr_frontend.services.get_translated_string(
             text=_('Welcome to DSMR-reader! Please make sure to check your settings in the Configuration page!')
         ),
-        redirect_to='frontend:configuration'
+        redirect_to='admin:index'
     )
 
     Notification.objects.create(