+++ dsmr_frontend/migrations/0008_merge_electricity_tariffs_notification.py
@@ -17,7 +17,7 @@ def insert_notifications(apps, schema_editor):
                 " frontend configuration: 'Merge electricity tariffs'"
             )
         ),
-        redirect_to='frontend:configuration'
+        redirect_to='admin:index'
     )