+++ dsmr_weather/tests/test_settings.py
@@ -13,6 +13,9 @@ class TestSettings(TestCase):
         """ Model should be registered in Django Admin. """
         self.assertTrue(site.is_registered(WeatherSettings))
 
+    def test_to_string(self):
+        self.assertNotEqual(str(self.instance), '{} object'.format(self.instance.__class__.__name__))
+
     def test_track(self):
         self.assertFalse(self.instance.track)