+++ docs/changelog.rst
@@ -8,6 +8,7 @@ v1.4.0 - 2016-xx-xx
 - Support for Python 3.5 (`#55 <https://github.com/dennissiemensma/dsmr-reader/issues/55>`_).
 - Support for both single and high/low tariff (`#130 <https://github.com/dennissiemensma/dsmr-reader/issues/130>`_).
 - Improve code coverage (`#151 <https://github.com/dennissiemensma/dsmr-reader/issues/151>`_).
+- Add new note from Dashboard has wrong time format (`#159 <https://github.com/dennissiemensma/dsmr-reader/issues/159>`_).
 
 
 v1.3.2 - 2016-11-08